Resolve review comments
- Removes list of permissions sent to channel
- Changes defense to load from specific channel id. You must setup env `DEFENSE_CATEGORY_ID` to have it function correctly.
- Merges into one function
